requirements.txt
pandas~=2.2.1
matplotlib~=3.8.3
PyYAML~=6.0.1
numpy~=1.26.4
xarray~=2024.2.0
shapely~=2.0.3
scipy~=1.12.0
pywake~=0.1.3
OrcFxAPI~=11.4.2
autograd~=1.6.2
tqdm~=4.66.2
setuptools~=65.5.1
pyproj~=3.6.1
optuna~=3.5.0
windrose~=1.9.0